Add link to run on MyBinder.org to run jupyter notebook for Free.
Hi Peter, thanks for putting this on GitHub, please find attached a
Patch that add :
  - a requirements.txt (that only add numpy so far)
  - and a link to launch on MyBinder.org.

MyBinder.org is a service that we (the Jupyter team) run for free, (no
login required, no Personal Information collected), that turn any GitHub
repository with Jupyter Notebooks into an ephemeral docker image which
is launched on Google Cloud Engine for a couple of hours. Images are
built on the fly if repository have changes, so it may take a few
minutes to launch after merged pull-request, or pushes.Otherwise it
should be quasi instantaneous, so the link added to the Readme should
alway be up to date (all of that is of course conditional to our servers
having enough capacity and we know you can be quite popular... So sorry
if you encounter delays)! MyBinder needs the requirement.txt to figure
out what to install (when running Python, because it can run R,
Julia...) so I added numpy, which seem necessary for some notebooks !

To anyone reading, the images are ephemeral so once you leave the tab,
your changes are gone and there is not way to save it ! Download
notebooks on your computer if you have significant modifications !

We (The Jupyter team) are a Big fans of what you are doing with Python
and Jupyter; and having your notebooks on GitHub is awesome. I hope
adding Binder to them would encourage people to play with it.

It's completely ok if you wish not to add the badge, it just requires
readers to past info manually on https://mybinder.org/, or use the
[chrome](https://chrome.google.com/webstore/detail/open-with-binder/napgohblobncpnagnehjpooinnimhpkc)
or
[firefox](https://addons.mozilla.org/en-US/firefox/addon/open-with-binder/)
extensions.

Thanks !
--
Matthias