Some minor style/reference changes
I attempted the indent the description of the different graphs so that the following paragraph about the conclusions doesn't look like it's a second paragraph about the last graph. I'm not sure it will render correctly because github is not clever enough about displaying the preview for changes to an ipython notebook.

I also changed the colors described since they don't match what I see on github but it's possible there's a color rendering difference that's going to be hard to pin down.

Lastly, I changed the number on the description of the simulation to match what's shown in the tables and graphs.